Move exception package to api
