+++ src/main/java/me/zeroeightsix/fiber/api/builder/ConfigAggregateBuilder.java
@@ -2,7 +2,7 @@ package me.zeroeightsix.fiber.api.builder;
 
 import me.zeroeightsix.fiber.api.annotation.AnnotatedSettings;
 import me.zeroeightsix.fiber.api.builder.constraint.AggregateConstraintsBuilder;
-import me.zeroeightsix.fiber.exception.RuntimeFiberException;
+import me.zeroeightsix.fiber.api.exception.RuntimeFiberException;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
