+++ src/main/java/me/zeroeightsix/fiber/api/builder/ConfigLeafBuilder.java
@@ -3,8 +3,8 @@ package me.zeroeightsix.fiber.api.builder;
 import me.zeroeightsix.fiber.api.builder.constraint.ConstraintsBuilder;
 import me.zeroeightsix.fiber.api.constraint.Constraint;
 import me.zeroeightsix.fiber.impl.constraint.FinalConstraint;
-import me.zeroeightsix.fiber.exception.FiberException;
-import me.zeroeightsix.fiber.exception.RuntimeFiberException;
+import me.zeroeightsix.fiber.api.exception.FiberException;
+import me.zeroeightsix.fiber.api.exception.RuntimeFiberException;
 import me.zeroeightsix.fiber.tree.ConfigLeaf;
 import me.zeroeightsix.fiber.tree.ConfigLeafImpl;
 import me.zeroeightsix.fiber.tree.ConfigTree;
