+++ src/main/java/me/zeroeightsix/fiber/api/builder/constraint/ComponentConstraintsBuilder.java
@@ -4,7 +4,7 @@ import me.zeroeightsix.fiber.api.builder.ConfigAggregateBuilder;
 import me.zeroeightsix.fiber.api.constraint.Constraint;
 import me.zeroeightsix.fiber.api.constraint.ConstraintType;
 import me.zeroeightsix.fiber.impl.constraint.ValuedConstraint;
-import me.zeroeightsix.fiber.exception.RuntimeFiberException;
+import me.zeroeightsix.fiber.api.exception.RuntimeFiberException;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
