+++ src/main/java/me/zeroeightsix/fiber/api/exception/FiberException.java
@@ -1,4 +1,4 @@
-package me.zeroeightsix.fiber.exception;
+package me.zeroeightsix.fiber.api.exception;
 
 /**
  * An exception thrown by fiber when something goes unrecoverably wrong. Unlike {@link RuntimeFiberException}, this exception must be caught.
