+++ src/main/java/me/zeroeightsix/fiber/impl/constraint/LengthConstraint.java
@@ -1,7 +1,7 @@
 package me.zeroeightsix.fiber.impl.constraint;
 
 import me.zeroeightsix.fiber.api.constraint.ConstraintType;
-import me.zeroeightsix.fiber.exception.RuntimeFiberException;
+import me.zeroeightsix.fiber.api.exception.RuntimeFiberException;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
