+++ src/main/java/me/zeroeightsix/fiber/serialization/JanksonSerializer.java
@@ -5,7 +5,7 @@ import blue.endless.jankson.JsonElement;
 import blue.endless.jankson.JsonObject;
 import blue.endless.jankson.api.SyntaxError;
 import me.zeroeightsix.fiber.Identifier;
-import me.zeroeightsix.fiber.exception.FiberException;
+import me.zeroeightsix.fiber.api.exception.FiberException;
 import me.zeroeightsix.fiber.tree.*;
 
 import javax.annotation.Nonnull;
