+++ src/test/java/me/zeroeightsix/fiber/tree/ConfigQueryTest.java
@@ -1,7 +1,7 @@
 package me.zeroeightsix.fiber.tree;
 
 import me.zeroeightsix.fiber.api.builder.ConfigTreeBuilder;
-import me.zeroeightsix.fiber.exception.FiberQueryException;
+import me.zeroeightsix.fiber.api.exception.FiberQueryException;
 import org.junit.jupiter.api.Test;
 
 import java.util.concurrent.atomic.AtomicReference;
